Extract StopViewHolder and TourAdapter
